backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
python-multipart
pymongo[srv]
emails
jinja2
pydantic[email]
pydantic-settings
python-dateutil
requests
email-validator